+++ packages/svelte/vitest.config.ts
@@ -1,0 +1,18 @@
+import { svelte } from '@sveltejs/vite-plugin-svelte';
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  plugins: [svelte()],
+  resolve: {
+    conditions: ['browser'],
+    extensions: ['.ts', '.tsx', '.js', '.jsx', '.mjs', '.cjs', '.json', '.svelte'],
+  },
+  ssr: {
+    noExternal: ['@mhaadi/svg'],
+  },
+  test: {
+    include: ['test/**/*.test.ts'],
+    environment: 'happy-dom',
+    setupFiles: ['./vitest.setup.ts'],
+  },
+});